Update version by changing a symlink
Previously, mattermost was never updated after its initial installation.
See #7.

Now we install mattermost in a separate folder per version. We check the
if the configured version is installed and if not, install it (*). The
newly installed version is then deployed by simply changing the symlink at
`/opt/mattermost`.

(*) In https://docs.mattermost.com/administration/upgrade.html it's suggested
to keep certain folders when updating, so we sync those from the current
version to the newly installed version.